backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for Clarifier"
authors = ["Max Boone <max.boone@umail.leidenuniv.nl>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1"
django-ninja = "^1.3.0"
PyJWT = "^2.9.0"
django-cors-headers = "^4.4.0"
daphne = "^4.1.2"
channels = "^4.1.0"
django-rq = "^2.10.2"
psycopg2-binary = "^2.9.9"
clickhouse-connect = "^0.8.3"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.3"
django-stubs = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.django-stubs]
django_settings_module = "backend.settings"
